+++ pkg/protocol/xprotocol/tars/protocol.go
@@ -124,5 +124,6 @@ func (proto tarsProtocol) EnableWorkerPool() bool {
 }
 
 func (proto tarsProtocol) GenerateRequestID(streamID *uint64) uint64 {
-	return atomic.AddUint64(streamID, 1)
+	// fix bug for issue: https://github.com/mosn/mosn/issues/2403
+	return uint64(int32(atomic.AddUint64(streamID, 1)))
 }
